move centerlines to separate module
